Refactors introduction check to dedicated module
Moves the logic for determining whether to play the introduction into a separate module for better organization and readability.
